ai-service/requirements.txt
fastapi
uvicorn
pydantic
pillow
exifread
numpy
scikit-learn
python-multipart
python-dotenv
pandas
joblib
geopy
opencv-python
imagehash
easyocr
scikit-learn
pandas
joblib
faker
python-dotenv
